Refactor process management and add new utility functions

+++ src/utils.c
@@ -2,10 +2,10 @@
 #include <stdlib.h>
 #include <time.h>
 #include <unistd.h>
-#include "utils.h"
+#include "../include/core/utils.h"
 
 void init_random(void) {
-    srand(time(NULL));
+    srand((unsigned int)time(NULL));
 }
 
 int random_range(int min, int max) {
@@ -16,7 +16,7 @@ void log_message(const char* message) {
     time_t now = time(NULL);
     char timestamp[26];
     ctime_r(&now, timestamp);
-    timestamp[24] = '\0'; // Remove newline
+    timestamp[24] = '\0';  // Remove newline
     printf("[%s] %s\n", timestamp, message);
 }
